requirements.txt
fastapi
uvicorn
langchain
langchain-groq
langchain-community
langchain-huggingface
chromadb
sentence-transformers
python-dotenv
requests
